requirements.txt
Pypdf2
python-dotenv
openai
flask
langchain
faiss-cpu
huggingface_hub
InstructorEmbedding
sentence_transformers
tiktoken